+++ frontend/src/app/(dashboard)/layout.module.scss
@@ -1,0 +1,3 @@
+.content {
+  display: flex;
+}
