+++ frontend/src/app/(dashboard)/member/page.module.scss
@@ -1,0 +1,162 @@
+@import url('https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@400;500;700&display=swap');
+
+/* 메인 레이아웃 */
+.layout {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  min-height: 100vh;
+  background-color: #f8f9fa;
+}
+
+/* 메인 콘텐츠 영역 */
+.mainContent {
+  padding: 32px;
+  display: flex;
+  flex-direction: column;
+}
+
+/* 카드 관련 스타일은 CardContainer 컴포넌트로 이동함 */
+
+.searchContainer {
+  position: relative;
+  width: 320px;
+}
+
+.searchInput {
+  padding: 12px 12px 12px 45px;
+  background-color: #FAFAFA;
+  border: 1px solid #E8E8E8;
+  border-radius: 10px;
+  width: 100%;
+  font-family: 'Noto Sans KR', sans-serif;
+  font-weight: 400;
+  font-size: 16px;
+  color: #333;
+}
+
+.searchInput::placeholder {
+  color: #6e6e6e;
+}
+
+.searchIcon {
+  position: absolute;
+  left: 15px;
+  top: 50%;
+  transform: translateY(-50%);
+  width: 17px;
+  height: 17px;
+  color: #6e6e6e;
+  pointer-events: none;
+}
+
+.createButton, .primaryButton {
+  text-align: center;
+}
+
+.createButton {
+  width: 110px;
+  height: 40px;
+  background-color: white;
+  color: #004bd6;
+  border: 1px solid #004bd6;
+  border-radius: 40px;
+  padding: 0;
+  cursor: pointer;
+  font-family: 'Noto Sans KR', sans-serif;
+  font-weight: 500;
+  font-size: 14px;
+  transition: all 0.2s ease-in-out;
+}
+
+.createButton:hover {
+  background-color: #f5f8ff;
+  color: #003cad;
+  border-color: #003cad;
+}
+
+.primaryButton {
+  width: 110px;
+  height: 40px;
+  background-color: #004bd6;
+  color: white;
+  border: none;
+  border-radius: 40px;
+  padding: 0;
+  cursor: pointer;
+  font-family: 'Noto Sans KR', sans-serif;
+  font-weight: 500;
+  font-size: 14px;
+  transition: all 0.2s ease-in-out;
+}
+
+.primaryButton:hover {
+  background-color: #003cad;
+}
+
+/* 테이블 */
+.tableContainer {
+  overflow: hidden;
+  margin-bottom: 0;
+  margin-top: 0;
+  border-top: 1px solid #eaeaea;
+  border-bottom: 1px solid #eaeaea;
+}
+
+.table {
+  width: 100%;
+  border-collapse: collapse;
+}
+
+.table th {
+  text-align: center;
+  padding: 11px 16px;
+  font-family: 'Noto Sans KR', sans-serif;
+  font-weight: 500;
+  font-size: 16px;
+  color: #000000;
+  background-color: transparent;
+  border-bottom: 1px solid #eaeaea;
+}
+
+.table td {
+  text-align: center;
+  padding: 11px 16px;
+  border-bottom: 1px solid #eaeaea;
+  font-family: 'Noto Sans KR', sans-serif;
+  font-weight: 400;
+  font-size: 16px;
+  color: #6a6a6a;
+}
+
+.checkboxColumn {
+  width: 40px;
+}
+
+.checkbox {
+  width: 20px;
+  height: 20px;
+  cursor: pointer;
+}
+
+.detailButton {
+  width: 69.051px;
+  height: 34px;
+  flex-shrink: 0;
+  background-color: white;
+  color: #6e6e6e;
+  border: 1px solid #6e6e6e;
+  border-radius: 4px;
+  padding: 0;
+  cursor: pointer;
+  font-family: 'Noto Sans KR', sans-serif;
+  font-weight: 500;
+  font-size: 14px;
+  transition: all 0.2s ease-in-out;
+}
+
+.detailButton:hover {
+  background-color: #f5f5f5;
+  color: #555555;
+  border-color: #555555;
+}
