--- frontend/src/app/components/Header/index.module.scss
@@ -1,41 +1,0 @@
-.header{
-  width: 100vw;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  height: 62px;
-  background: #FFFFFF;
-  box-shadow: 0 2px 4px 0 #00000026;
-  padding: 0 54px;
-
-  font-family: Noto Sans KR;
-  font-weight: 400;
-  font-size: 16px;
-  color: var(--black);
-}
-
-.header-container {
-  display: flex;
-  align-items: center;
-  width: 100%;
-
-  .header-items-wrapper {
-    display: flex;
-    align-items: center;
-    margin-left: 100px;
-    gap: 110px;
-  }
-
-  .header-auth-items-wrapper {
-    display: flex;
-    align-items: center;
-    margin-left: auto;
-    gap: 32px;
-
-    color: var(--gray-2);
-  }
-}
-
-.strong {
-  color: var(--blue);
-}
