--- frontend/src/app/components/Input/ImageUploader/index.module.scss
@@ -1,80 +1,0 @@
-.uploadContainer {
-  display: flex;
-  flex: 1;
-  justify-content: center;
-  align-items: center;
-  border: 1px dashed var(--gray-3);
-  border-radius: 20px;
-  background-color: var(--gray-6);
-  cursor: pointer;
-}
-
-.uploadLabel {
-  display: flex;
-  width: 100%;
-  height: 100%;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-
-  .uploadBox {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 20px;
-
-    p {
-      font-family: Noto Sans KR;
-      font-weight: 400;
-      font-size: 14px;
-      line-height: 18px;
-      color: var(--gray-2);
-    }
-  }
-}
-
-.uploading {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 20px;
-
-  font-family: Noto Sans KR;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 18px;
-  color: var(--black);
-
-
-  .progressContainer {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    position: relative;
-    width: 68px;
-    height: 68px;
-
-    font-family: Pretendard;
-    font-weight: 500;
-    font-size: 12px;
-    line-height: 18px;
-    color: var(--primary-blue);
-
-    svg {
-      position: absolute;
-      top: 0;
-      left: 0;
-      width: 68px;
-      height: 68px;
-    }
-  }
-}
-
-.preview img {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  border-radius: 20px;
-}
