--- frontend/src/app/components/Input/TextInput/index.module.scss
@@ -1,38 +1,0 @@
-.container {
-  display: flex;
-  flex: 1;
-  width: 100%;
-  flex-direction: column;
-  gap: 12px;
-}
-
-.label {
-  font-family: Noto Sans KR;
-  font-weight: 500;
-  font-size: 18px;
-  line-height: 18px;
-
-  color: var(--black); // 진한 검은색
-}
-
-.input {
-  width: 100%;
-  padding: 11px 10px;
-  border: 1px solid var(--gray-8);
-  border-radius: 8px;
-  background-color: var(--gray-6);
-  outline: none;
-
-  font-family: Noto Sans KR;
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 18px;
-
-  &::placeholder {
-    color: var(--gray-7);
-  }
-
-  &:focus {
-    //border-color: #;
-  }
-}
