+++ frontend/src/components/CardContainer/index.module.scss
@@ -2,6 +2,9 @@
 
 /* 카드 */
 .card {
+  display: flex;
+  flex-direction: column;
+  flex: 1;
   background-color: white;
   border-radius: 40px;
   padding: 52px;
