--- frontend/src/components/Header/app/layout.js
@@ -1,18 +1,0 @@
-import "./styles/globals.scss";
-import Header from "@/components/Header";
-
-export const metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
-};
-
-export default function RootLayout({ children }) {
-  return (
-    <html lang="en">
-      <body>
-        <Header />
-        {children}
-      </body>
-    </html>
-  );
-}
