--- frontend/src/components/Header/app/page.module.css
@@ -1,59 +1,0 @@
-.container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  min-height: 100vh;
-  background-color: #f6f7fb;
-}
-
-.card {
-  background-color: #fff;
-  padding: 60px 63px;
-  border-radius: 20px;
-  box-shadow: 0 5px 30px rgba(0, 0, 0, 0.15);
-  width: 500px;
-  height: 520px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.logo {
-  width: 153px;
-  height: 27.2px;
-  margin-bottom: 54px;
-}
-
-.form {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-}
-
-/* 에러 메시지 스타일 */
-.errorMessage {
-  font-family: 'Noto Sans KR', sans-serif;
-  font-weight: 400; /* Regular */
-  font-size: 8px;
-  color: #DF373C;
-  height: 30px;
-  line-height: 30px; /* 수직 중앙 정렬 */
-  text-align: center;
-  margin-bottom: 20px;
-}
-
-.loginButton {
-  padding: 14px;
-  background-color: #004bd6;
-  color: #fff;
-  border: none;
-  border-radius: 10px;
-  cursor: pointer;
-  font-size: 18px;
-  font-weight: 500;
-  transition: background-color 0.2s;
-
-  &:hover {
-    background-color: #003bb5;
-  }
-}
