--- frontend/src/components/Header/app/styles/colors.scss
@@ -1,19 +1,0 @@
-:root {
-  // Primary Colors
-  --primary-blue: #004BD6;
-  --primary-cyan: #15FFEB;
-  --primary-red: #DF373C;
-
-  // Neutral Colors
-  --black: #000000;
-  --white: #FFFFFF;
-
-  // Gray Scale
-  --gray-1: #545F71;
-  --gray-2: #95989D;
-  --gray-3: #D9D9D9;
-  --gray-4: #CDCDCD;
-  --gray-5: #F4F5F8;
-  --gray-6: #FAFAFA;
-  --gray-transparent: #00000033;
-}
