--- frontend/src/components/Header/app/styles/fonts.scss
@@ -1,8 +1,0 @@
-:root {
-  // EXAMPLE
-  //  @mixin Noto-sans($size, $weight) {
-  //    font-family: Noto Sans KR;
-  //    font-weight: var($weight);
-  //    font-size: var($size);
-  //  }
-}
