--- frontend/src/components/Header/app/styles/globals.scss
@@ -1,33 +1,0 @@
-@import "./colors.scss";
-@import "./fonts.scss";
-//@import "./mixins.scss";
-
-// 기본 스타일 설정
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-// 기본 배경 및 글꼴 설정
-html, body {
-  max-width: 100vw;
-  overflow-x: hidden;
-  height: 100%;
-  background: var(--gray-5);
-}
-
-// Next.js 기본 링크 스타일
-a {
-  text-decoration: none;
-  color: inherit;
-}
-
-button {
-  all: unset; // 기본 버튼 스타일 제거
-  cursor: pointer;
-}
-
-ul, ol {
-  list-style: none;
-}
