--- frontend/src/components/Header/components/Button/LogoutButton/index.jsx
@@ -1,19 +1,0 @@
-import Button from "../";
-
-const LogoutButton = ({ onLogout }) => {
-    const handleLogout = () => {
-        if (onLogout) {
-            onLogout();
-        } else {
-            console.log("로그아웃");
-        }
-    };
-
-    return (
-        <Button variant="outline" onClick={handleLogout}>
-            로그아웃
-        </Button>
-    );
-};
-
-export default LogoutButton;
