--- frontend/src/components/Header/components/Button/index.module.scss
@@ -1,34 +1,0 @@
-.button {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 6px 22px;
-  border-radius: 50px;
-  cursor: pointer;
-
-  font-family: Noto Sans KR;
-  font-weight: 400;
-  font-size: 12px;
-  line-height: 18px;
-  color: var(--gray-2);
-
-  &:hover {
-    //background-color: ;
-  }
-
-  &:active {
-    //background-color: ;
-  }
-}
-
-.outline {
-  border: 1px solid var(--gray-2);
-  background-color: transparent;
-  color: var(--gray-2);
-}
-
-.default {
-  background: 1px solid var(--primary-blue);
-  border: 1px solid var(--primary-blue);
-  color: var(--white);
-}
